+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Default value for `display` prop in `PlaceDetails` component.
 - Prop `hiddenFields` in `AddressForm` component.
 - Prop `onNoPostalCode` in `LocationInput` component.
+- Prop `variation` in `LocationInput` component.
 
 ### Changed
 - Update design of `LocationInput` component to match their spec.